requirements.txt
scrapy==2.13.3
pymysql==1.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
sqlalchemy==2.0.35
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
fake-useragent==1.5.1
python-decouple==3.8